requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.2.0
executing==2.2.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
icecream==2.1.4
idna==3.10
loguru==0.7.3
motor==3.7.0
passlib==1.7.4
pip==24.2
psutil==7.0.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pygments==2.19.1
pymongo==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
reactivex==4.0.4
rsa==4.9.1
setuptools==79.0.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
